requirements.txt
anyio==3.6.2
astroid==2.14.2
attrs==22.2.0
build==0.10.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
coverage==7.2.1
dill==0.3.6
flake8==6.0.0
ghp-import==2.1.0
griffe==0.25.5
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
Markdown==3.3.7
MarkupSafe==2.1.2
mccabe==0.7.0
mergedeep==1.3.4
mkdocs==1.4.2
mkdocs-autorefs==0.4.1
mkdocs-material==9.1.0
mkdocs-material-extensions==1.1.1
mkdocstrings==0.20.0
mkdocstrings-python==0.8.3
packaging==23.0
platformdirs==3.1.0
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.14.0
pylint==2.16.3
pymdown-extensions==9.9.2
pyproject_hooks==1.0.0
pytest==7.2.2
python-dateutil==2.8.2
python-telegram-bot==20.1
PyYAML==6.0
pyyaml_env_tag==0.1
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tomlkit==0.11.6
urllib3==1.26.14
watchdog==2.3.1
wrapt==1.15.0
